[YS-447] refactor: 컴포넌트명 변경 및 파일 위치 수정
