+++ src/app/post/[post_id]/mobile/components/ExperimentPostMobileContainer/ExperimentPostMobileContainer.tsx
@@ -1,117 +1,128 @@
 'use client';
 
 import * as Toast from '@radix-ui/react-toast';
+import { useParams, useRouter } from 'next/navigation';
 import { useState } from 'react';
 
-import {
-  buttonGradientBackground,
-  contactButton,
-  emptyView,
-  emptyViewTitle,
-  experimentPostMobileContainerLayout,
-  fixedBottomButtonLayout,
-} from './ExperimentPostMobileContainer.css';
-import useExperimentDetailsQuery from '../../../hooks/useExperimentDetailsQuery';
-import ExperimentPostInfo from '../ExperimentPostInfo/ExperimentPostInfo';
-import ExperimentPostTabs from '../ExperimentPostTabs/ExperimentPostTabs';
-import ParticipationGuideBottomSheet from '../ParticipationGuideBottomSheet/ParticipationGuideBottomSheet';
+import ExperimentPostMobileContainer from '../ExperimentPostMobileContentContainer/ExperimentPostMobileContentContainer';
+import ExperimentPostMobileHeader from '../ExperimentPostMobileHeader/ExperimentPostMobileHeader';
 import {
   copyToastLayout,
   copyToastTitle,
   copyToastViewport,
 } from '../ParticipationGuideBottomSheet/ParticipationGuideBottomSheet.css';
+import PostMenuBottomSheet from '../PostMenuBottomSheet/PostMenuBottomSheet';
 
-import Button from '@/components/Button/Button';
+import useDeleteExperimentPostMutation from '@/app/my-posts/hooks/useDeleteExperimentPostMutation';
 import Icon from '@/components/Icon';
+import ConfirmModal from '@/components/Modal/ConfirmModal/ConfirmModal';
 import useOverlay from '@/hooks/useOverlay';
 import { colors } from '@/styles/colors';
-import Spinner from '@/components/Spinner/Spinner';
-import { getErrorMessage } from '../../../ExperimentPostPage.utils';
 
-const ExperimentPostMobileContainer = ({
-  experimentDetailResponse,
-  postId,
-}: {
-  experimentDetailResponse: ReturnType<typeof useExperimentDetailsQuery>;
-  postId: string;
-}) => {
+import EditNotReadyModal from '../EditNotReadyModal/EditNotReadyModal';
+import useExperimentDetailsQuery from '../../../hooks/useExperimentDetailsQuery';
+
+const ExperimentPostMobileContentContainer = () => {
   const { open, close } = useOverlay();
-  const [isCopyToastOpen, setIsCopyToastOpen] = useState(false);
+  const [isEditModalOpen, setIsEditModalOpen] = useState(false);
+  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
+  const [isToastOpen, setIsToastOpen] = useState(false);
+  const [toastMessage, setToastMessage] = useState('');
 
-  const postDetailData = experimentDetailResponse.data;
+  const { post_id } = useParams();
+  const postId = Array.isArray(post_id) ? post_id[0] : post_id;
+  const router = useRouter();
 
-  if (experimentDetailResponse.isLoading) {
-    return (
-      <div className={emptyView}>
-        <Spinner />;
-      </div>
+  /* 특정 공고 상세 조회 */
+  const experimentDetailResponse = useExperimentDetailsQuery({ postId });
+
+  /* 공고 삭제 */
+  const { mutate: deleteExperimentPostMutation } = useDeleteExperimentPostMutation();
+
+  const handleDeletePost = () => {
+    setIsDeleteModalOpen(false);
+    setToastMessage('공고를 삭제하였습니다.');
+    setIsToastOpen(true);
+
+    deleteExperimentPostMutation(
+      { postId },
+      {
+        onSuccess: () => {
+          setTimeout(() => {
+            router.push('/');
+            setToastMessage('');
+          }, 1700);
+        },
+        onError: () => {
+          setToastMessage('공고 삭제를 실패하였습니다. 잠시 후 다시 시도해주세요.');
+          setIsToastOpen(true);
+        },
+      },
     );
-  }
+  };
 
-  if (experimentDetailResponse.isError) {
-    return (
-      <div className={emptyView}>
-        <p className={emptyViewTitle}>{getErrorMessage(experimentDetailResponse.error)}</p>
-        <button onClick={() => experimentDetailResponse.refetch()} className={contactButton}>
-          재시도
-        </button>
-      </div>
-    );
-  }
-
-  if (!postDetailData)
-    return (
-      <div className={emptyView}>
-        <div className={emptyViewTitle}>공고 상세 정보가 없습니다.</div>
-      </div>
-    );
-
-  const handleOpenBottomSheet = () => {
+  const handleOpenMenuBottomSheet = () => {
     open(
       () => (
-        <ParticipationGuideBottomSheet
-          onConfirm={close}
+        <PostMenuBottomSheet
           postId={postId}
-          setIsToastOpen={setIsCopyToastOpen}
+          onConfirm={close}
+          onEditClick={() => setIsEditModalOpen(true)}
+          onDeleteClick={() => setIsDeleteModalOpen(true)}
         />
       ),
       {
-        title: '참여 방법',
-        headerMode: 'none',
+        headerMode: 'drag-handle',
       },
     );
   };
 
   return (
-    <div className={experimentPostMobileContainerLayout}>
-      <ExperimentPostInfo postDetailData={postDetailData} />
-      <ExperimentPostTabs postDetailData={postDetailData} />
-
-      <div className={buttonGradientBackground}>
-        <div className={fixedBottomButtonLayout}>
-          <Button variant="dark" size="medium" height={'5.6rem'} onClick={handleOpenBottomSheet}>
-            참여 방법 확인하기
-          </Button>
-        </div>
+    <>
+      <div>
+        <ExperimentPostMobileHeader
+          onOpenMenuBottomSheet={handleOpenMenuBottomSheet}
+          experimentDetailResponse={experimentDetailResponse}
+        />
+        <ExperimentPostMobileContainer
+          experimentDetailResponse={experimentDetailResponse}
+          postId={postId}
+        />
       </div>
 
-      {/* 복사 성공 토스트 알림 */}
+      {/* 공고 수정 모바일 화면 준비중 모달 */}
+      <EditNotReadyModal isOpen={isEditModalOpen} onOpenChange={setIsEditModalOpen} />
+
+      {/* 삭제 확인 모달 */}
+      <ConfirmModal
+        isOpen={isDeleteModalOpen}
+        onOpenChange={setIsDeleteModalOpen}
+        confirmTitle="해당 공고를 삭제할까요?"
+        cancelText="닫기"
+        confirmText="삭제하기"
+        confirmButtonColor={colors.field09}
+        onConfirm={handleDeletePost}
+        isMobile
+        closeIcon={false}
+      />
+
+      {/* 삭제 성공 토스트 알림 */}
       <Toast.Provider swipeDirection="right">
         <Toast.Root
           className={copyToastLayout}
-          open={isCopyToastOpen}
-          onOpenChange={setIsCopyToastOpen}
-          duration={1500}
+          open={isToastOpen}
+          onOpenChange={setIsToastOpen}
+          duration={1700}
         >
           <Toast.Title className={copyToastTitle}>
-            <Icon icon="CheckRound" color={colors.primaryMint} width={24} height={24} />
-            <p>복사되었어요</p>
+            <Icon icon="Alert" color={colors.textAlert} width={24} height={24} />
+            <p>{toastMessage}</p>
           </Toast.Title>
         </Toast.Root>
         <Toast.Viewport className={copyToastViewport} />
       </Toast.Provider>
-    </div>
+    </>
   );
 };
 
-export default ExperimentPostMobileContainer;
+export default ExperimentPostMobileContentContainer;
