+++ src/app/post/[post_id]/mobile/components/ExperimentPostMobileContentContainer/ExperimentPostMobileContentContainer.css.ts
@@ -3,7 +3,7 @@ import { style } from '@vanilla-extract/css';
 import { colors } from '@/styles/colors';
 import { fonts } from '@/styles/fonts.css';
 
-export const experimentPostMobileContainerLayout = style({
+export const experimentPostMobileContentContainerLayout = style({
   minHeight: 'calc(100dvh - 5.4rem)',
   backgroundColor: colors.field01,
